+++ docs/about/history.rst
@@ -32,6 +32,7 @@ Changed
    -  Remove outdated lighting validation.
    -  Validate layers to warn users of potential build failures if *Crest* related renderers are not on the same layer as the *OceanRenderer.Layer*.
    -  No longer log info level validation to the console.
+   -  Add info validation for tips on using reflection probes when found in a scene.
 
 
 Fixed